New vip-spec.sv VIP package specification file
Experiment in specfying the files and arguments for Bathtub in a SystemVerilog native file format.
The motivation is that other programming languages have well-defined native means for managing packages.
For example, Node.js has `package.json` and Python has `__init__.py`.
This Bathtub file specifies its files and metadata with a native SystemVerilog struct variable, `spec`.
Run this file as a program in a simulator.
It outputs a ".f" argument file into the current working directory.
Use the argument file in your simulations.