Prepare UVM codec example and test.
UVM 1.0, 1.1, and and 1.2 reference implementation libraries include an "integrated" example called "codec."
This example will add Bathtub on the codec testbench.
The first step is to ensure we can run the codec example as-is.

We copy the entire UVM 1.2 library into a personal working directory.
We tweak a couple makefiles, then run the codec test in the working directory.

Why copy the entire directory?
The codec example, as delivered, is self-contained and runs inside the UVM library directories with relative paths, etc.
So it's simplest to stick with that strategy.

- new file:   examples/uvm_examples/README.md
- new file:   examples/uvm_examples/codec/README.md
  - Placeholders

- modified:   test/conftest.py
  - New fixture `get_test_config()` gives tests access to our test config object, which lists local paths to simulator and UVM library installations.

- new file:   test/examples/__init__.py
- new file:   test/examples/uvm_examples/__init__.py
- new file:   test/examples/uvm_examples/codec/__init__.py
  - Declare Python packages to prevent namespace collisions.

- new file:   test/examples/uvm_examples/codec/Makefile_inc.xcelium
- new file:   test/examples/uvm_examples/codec/Makefile_run.xcelium
  - New Makefiles for running Xcelium simulations.
  - UVM reference library only supports older Incisive (ius), so we add Xcelium support.
  - Makefiles are based on the existing Makefile.ius files.

- new file:   test/examples/uvm_examples/codec/codec_test.py
  - New pytest module.
  - First test `test_codec_as_is()` runs the UVM 1.2 codec example in Xcelium as-is.
  - Passes.