Clean up how default sequences are set.
Default sequences are set via the `uvm_config_db` interface.
Overriding the config db entries was fragile.
One elusive issue is that the `env` instance is _not_ a child of the `test` class; it's a child of `$root`.
This was causing me to use the wrong context.
Then there was the issue of setting `default_sequence` via instance vs. wrapper; see UVM documentation of `uvm_sequencer_base::start_phase_sequence()`.
Then it wasn't clear if setting the `default_sequence` resource in the base class and child class caused the resources to accumulate or overwrite.
If the lookup key is identical, it overwrites.
In debugging all this, `uvm_resource_db::dump()` was very helpful.

To understand the final resolution, compare the `build_phase()` methods in `test.sv` (base class) and `bathtub_test.svh` (child class).

- modified:   examples/uvm_examples/codec/Makefile_run.xcelium
  - New test target `test_with_bathtub` runs the original test on the modified testbench for sanity.
  - This test is for development only, is not included in the `all` target.

- modified:   examples/uvm_examples/codec/bathtub_test.svh
  - Set the default sequence in the build phase instead of the main phase.
  - We override the defaults set in the base test class.

- modified:   examples/uvm_examples/codec/tb_env.svh
  - Cleanup comments.

- new file:   examples/uvm_examples/codec/test.sv
  - Modify the test base class to set default sequences in the build phase.
  - Default sequences in the unmodified testbench are set by the `tb_env` class.

- modified:   test/examples/uvm_examples/codec/codec_test.py
  - Copy the modified `test` base class into the run directory.
  - Parameterize `test_codec_with_bathtub()` to run two different UVM tests.