Remove unneeded modifications to files.
The fewer files that need to be modified, the simpler this example becomes.
I realized I didn't need to modify `tb_env.svh` to add a virtual sequencer; I could add the virtual sequencer to my `bathtub_test.svh` instead.
Strictly speaking, `tb_env.svh` is the better place for the virtual sequencer, but for the purpose of this example, simplicity prevails.

- modified:   examples/uvm_examples/codec/bathtub_test.svh
- modified:   examples/uvm_examples/codec/uvm-1.0p1/bathtub_test.svh
  - Instantiate virtual sequencer here, instead of in `env`.
  - Connect the virtual sequencer in the `start_of_simulation` phase, because that's when the base class assigns the `env` member.

- deleted:    examples/uvm_examples/codec/tb_env.svh
- deleted:    examples/uvm_examples/codec/test.sv
- deleted:    examples/uvm_examples/codec/uvm-1.0p1/tb_env.svh
- deleted:    examples/uvm_examples/codec/uvm-1.0p1/test.sv
  - With the above change, these files don't need to be touched, so delete my modified versions.

- modified:   test/examples/uvm_examples/codec/codec_test.py
  - No need to copy over the unneeded modifed files anymore.