Add doc strings to methods.
After some deliberation, decided to add doc strings as SystemVerilog attributes attached to class methods.

It's a Verilogical solution:
- Attributes are part of the language
- They create strong binding between the doc string and the method
- They could eventually be used for introspection, with some DPI work
- They can be parsed and extracted by an external SystemVerilog parser
- They can be extracted reasonably by regular expression, which is probably the first thing I'll do

e.g.
```
sed -ne '/(\* *doc\([a-zA-Z_\$]*\) * =/,/;$/p' bathtub.svh
```

Drawbacks:
- SystemVerilog multi-line strings are ugly
- They could be beautiful with triple quotes, but my simulator versions don't support them yet, and I'm worried others might be in the same situation

In this implementation, I named the doc string attributes `doc$markdown`.
The idea is that other formats could be supported, such as `doc$naturaldocs` or `doc$javadoc`.
I opted not to include newlines in the doc strings.
Since I'll be extracting them with an external tool, I can preserve the newlines in that tool.

- modified:   src/bathtub_pkg/bathtub.svh
  - Added doc strings to `new()` and `configure()`.
  - Removed `report_object` from `configure()` and moved it to its own accessor.
    - It didn't fit in with the other parameters to `configure()`.