Remove set_step_attributes() from step definition interface.
Upon reviewing the step definition interface, I didn't want to expose the `set_step_attributes()` method to the user.
That is strictly for use by the runner in creating the step definition object, not for use inside the step definition object.
But if `set_step_attributes()` is removed, how can the runner provide the step definition object with its step attributes object?
The solution here is to use a UVM global pool.
I added a typedef to `bathtub_pkg`:
```sv
typedef uvm_pool#(uvm_sequence_base, step_nurture) step_attributes_pool_t;
```
When the runner creates a step definition object and a step attributes object, it stores the attributes object in the global pool with the definition object as the lookup key.
When the step definition's `get_step_attributes()` method is called, it retrieves the attributes from the global pool, and deletes the attributes from the global pool to keep the pool tidy.

This is not ideal because the global pool is, well, global.
Anyone could mess with the global pool.
But that's a problem for another day.

As of now, the objective of removing set_step_attrubutes() has been achieved.
The `step_definition` interface is now immutable; it has no setters.

- modified:   src/bathtub_macros.sv
  - Removed `set_step_attributes()` implementation as it's no longer required.
  - Moved the `__step_static_attributes` member into `get_step_attributes()` as a static variable, so not even user's step definition code can reach it.
  - `get_step_attributes()` retrieves the attributes object from the global pool, then deletes it from the pool.
  - Added checks and warnings in case the attributes object is not found in the pool or is null.

- modified:   src/bathtub_pkg/bathtub_pkg.svh
  - New typedef `step_attributes_pool_t` for the global step attributes pool.

- modified:   src/bathtub_pkg/gherkin_document_runner/gherkin_document_runner.svh
  - Pass step attributes to the step definition object indirectly through the global pool, not directly.

- modified:   src/bathtub_pkg/step_definition_interface.svh
  - Removed `set_step_attributes()` from the interface as it's no longer desired.