Add step accessors to step_definition_interface.
Upon reflection, I'm rethinking my earlier decision to push step definition functions into the first line of attributes and context sequence objects.
The primary purpose of the step_definition_interface is to be useful to the step definition writer.
Therefore it shall be a facade that contains as many useful methods as reasonable.
The implementations will delegate to the sub-objects as necessary, hiding that complexity from the user.

- modified:   src/bathtub_macros.sv
  - Implemented step accessors in step definition and virtual step definition macros.
  - Removed vestigial setters from virtual step definition macro.

- modified:   src/bathtub_pkg/step_definition_interface.svh
  - Cleaned up model. #WIP
  - New step accessors
    - `get_step_keyword()`
    - `get_step_text()`
    - `get_step_argument_data_table()`
    - `get_step_argument_doc_string()`

- modified:   src/bathtub_pkg/step_definition_seq.svh
  - Implement new step accessors.