Change bathtub from a uvm_object to a uvm_component.
After careful consideration, this was the correct decision.
Bathtub is rightfully a persistent component.
That required all its instantiations in all tests to change, to include the parent argument.
Updated docstrings.
The primary motiviation was that I wanted the bathtub object to be locatable in the component hierarchy.
If a `test_sequence` is created outside bathtub, for example by a test or environment, and the sequence is not configured with a `bathtub` reference, I want the sequence to be able to search the component hierarchy for potential bathtubs.

Up until now bathtub was an object because I felt it was closer to a sequence than a sequencer.
But I've changed my mind.

- modified:   src/bathtub_pkg/bathtub.svh
  - Changed base class from `uvm_report_object` to `uvm_component`.
  - Note that `uvm_component` is a `uvm_report_object`.
  - Updated the constructor to add the `parent` argument.
  - Left the `parent` argument without a default for now to flush out any existing tests that don't provide it.

- modified:   examples/alu_division/result/testbench.sv
- modified:   examples/uvm_examples/codec/bathtub_test.svh
- modified:   examples/uvm_examples/codec/uvm-1.0p1/bathtub_test.svh
- modified:   test/bathtub_pkg_test_suite/bathtub_unit_test.sv
- modified:   test/e2e/basic/tests/basic_test.svh
- modified:   test/e2e/gherkin_parser/tests/gherkin_parser_rules_test.svh
- modified:   test/e2e/plusargs/tests/plusarg_bathtub_dryrun_test.svh
- modified:   test/e2e/plusargs/tests/plusarg_bathtub_features_test.svh
- modified:   test/e2e/plusargs/tests/plusarg_bathtub_help_test.svh
- modified:   test/e2e/plusargs/tests/plusarg_bathtub_include_exclude_test.svh
- modified:   test/e2e/plusargs/tests/plusarg_bathtub_start_stop_test.svh
- modified:   test/e2e/plusargs/tests/plusarg_bathtub_verbosity_test.svh
- modified:   test/e2e/step_definition_interface/tests/simple_test.svh

- modified:   src/bathtub_pkg/step_definition_interface.svh
  - Edited docstrings to fix Markdown for inline code macros.