Implement new virtual sequencer concept.
After rethinking the idea of a `uvm_sequencer` virtual sequencer, formulated a new concept.
Treat the virtual sequence as a connector/adapter class between the `uvm_env` testbench object and the `uvm_sequence` virtual sequence object.
The `uvm_test` test is responsible for configuring the virtual sequencer with the `uvm_env` instance.
The `uvm_sequence` class comes with a "required interface" interface class, which provides all the sequence's dependencies.
The virtual sequencer is responsible for implementing the sequence required interface methods by retrieving them from the `uvm_env` instance.
The virtual sequencer is an adapter, a wrapper around the environment that provides the sequence's required interface.
This decouples the testbench from the sequences and clarifies reuse.
If a new environment is created--e.g., going from block-level to chip-level--then a new virtual sequencer class is required.
If sequences need a new resource, it must be added to the required interface, and retroactively implemented by existing virtual sequencers.

In UML ball-and-socket (or lollipop-and-socket) Component Diagrams, the virtual sequencer provides the ball/lollipop and the virtual sequence presents the socket.

I parameterized the virtual sequencer with the types of the environment and sequence required interface, as living documentation of what interfaces the class is adapting.
Similarly I parameterized the virtual sequence with the type of its required interface as living documentation.

This has some nice semantics.
- The virtual sequencer is the decoupling adapter between the environment and the virtual sequence.
- The virtual sequencer provides the interface which the virtual sequence requires.
- The virtual sequence's `p_sequencer` is the means of dependency injection into the sequence.
- The virtual sequence doesn't really need to be a sequence.
  As long as it implements the required interface, it could be any base class.
- With all the parameters and virtual functions, the test, environment, virtual sequencer, and virtual sequences can all be sub-classed with great flexibility.

- modified:   examples/uvm_examples/codec/bathtub_test.svh
- modified:   examples/uvm_examples/codec/uvm-1.0p1/bathtub_test.svh
  - Added type parameters to the virtual sequencer instantiation.
  - Configured the virtual sequencer with the environment, not the individual components.

- modified:   examples/uvm_examples/codec/codec_step_definitions.svh
- modified:   examples/uvm_examples/codec/uvm-1.0p1/codec_step_definitions.svh
  - Added a type parameter for the required interface--the socket--to the virtual sequence base class.
  - Added parentheses to function calls, even when there are no arguments, because they are now required by SystemVerilog.

- modified:   examples/uvm_examples/codec/tb_virtual_sequencer.svh
  - New required interface class `codec_step_definition_rif` for the testbench's virtual sequences.
    - The interface class is named after the virtual sequence base class.
    - The suffix "_rif" stands for "required interface."
    - Replaced individual resources from the environment with a single reference to the environment itself.
    - Added a getter and setter for the environment handle.
  - Added getters for the individual resources, implementing the required interface.